+++ R/ppc-loo.R
@@ -546,12 +546,12 @@ ppc_loo_intervals <-
            prob = 0.5,
            prob_outer = 0.9,
            alpha = 0.33,
-           size = 2.5,
+           size = 1,
            fatten = deprecated(),
            linewidth = 1,
            order = c("index", "median")) {
     check_ignored_arguments(..., ok_args = list("moment_match"))
-    size <- resolve_fatten(fatten, size, default_size = 2.5,
+    size <- resolve_fatten(fatten, size, default_size = 1,
                            calling_fn = "ppc_loo_intervals")
     y <- validate_y(y)
     order_by_median <- match.arg(order) == "median"