+++ R/ppd-intervals.R
@@ -57,7 +57,7 @@ ppd_intervals <-
            prob = 0.5,
            prob_outer = 0.9,
            alpha = 0.33,
-           size = 2.5,
+           size = 1,
            fatten = deprecated(),
            linewidth = 1) {
 
@@ -66,7 +66,7 @@ ppd_intervals <-
       check_ignored_arguments(...)
       dots$group <- NULL
     }
-    size <- resolve_fatten(fatten, size, default_size = 2.5,
+    size <- resolve_fatten(fatten, size, default_size = 1,
                            calling_fn = "ppd_intervals")
 
     data <- ppd_intervals_data(
@@ -111,7 +111,7 @@ ppd_intervals_grouped <-
            prob = 0.5,
            prob_outer = 0.9,
            alpha = 0.33,
-           size = 2.5,
+           size = 1,
            fatten = deprecated(),
            linewidth = 1) {
     check_ignored_arguments(...)